pyproject.toml
[project]
name = "demo-black-env"
version = "0.1.0"
description = ""
authors = [
    {name = "Thijs Meeuwisse",email = "thijs@meeuwis.se"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "black (>=26.1.0,<27.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "snowflake-connector-python[pandas, secure-local-storage] (>=4.0.0,<5.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false